Introduce switch entities

+++ components/daly_bms_ble/switch/daly_switch.cpp
@@ -9,7 +9,7 @@ static const char *const TAG = "daly_bms_ble.switch";
 
 void DalySwitch::dump_config() { LOG_SWITCH("", "DalyBmsBle Switch", this); }
 void DalySwitch::write_state(bool state) {
-  if (this->parent_->send_command(this->holding_register_, state)) {
+  if (this->parent_->send_command(this->holding_register_, (uint16_t) state)) {
     this->publish_state(state);
   }
 }
